pyproject.toml
[project]
name = "rpl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "fastapi>=0.121.1",
    "google-genai>=1.52.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.23.1",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
